+++ src/IR/logger/LLVMLogger.cpp
@@ -129,7 +129,7 @@ void LLVMLogger::setCurrentSourceLocation(
   _location = location;
 }
 
-unsigned int LLVMLogger::getErrorCount() const { return errorCount; }
+uint64_t LLVMLogger::getErrorCount() const { return errorCount; }
 
 DiagnosticUtils::SourceLocation LLVMLogger::getCurrentSourceLocation() const {
   return _location;
